+++ src/main/java/com/scottkillen/mod/koresample/TheMod.java
@@ -3,6 +3,7 @@ package com.scottkillen.mod.koresample;
 import cpw.mods.fml.common.Mod;
 import cpw.mods.fml.common.Mod.EventHandler;
 import cpw.mods.fml.common.event.FMLInitializationEvent;
+import cpw.mods.fml.common.event.FMLInterModComms;
 import cpw.mods.fml.common.event.FMLPostInitializationEvent;
 import cpw.mods.fml.common.event.FMLPreInitializationEvent;
 
@@ -13,11 +14,16 @@ public class TheMod
     public static final String MOD_NAME = "Kore Sample";
     public static final String MOD_VERSION = "${mod_version}";
     private static final String RESOURCE_PREFIX = MOD_ID.toLowerCase() + ':';
+    private static final String VERSIONCHECK_URL =
+            "https://raw.githubusercontent.com/ScottKillen/glowing-ninja/master/KoreSample.json";
 
     public static String resourcePrefix() { return RESOURCE_PREFIX; }
 
     @EventHandler
-    public void onFMLPreInitialization(FMLPreInitializationEvent event) { }
+    public void onFMLPreInitialization(FMLPreInitializationEvent event)
+    {
+        FMLInterModComms.sendRuntimeMessage(MOD_ID, "VersionChecker", "addVersionCheck", VERSIONCHECK_URL);
+    }
 
     @EventHandler
     public void onFMLInitialization(FMLInitializationEvent event) { }