+++ src/main/resources/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+- Add support for [Version Checker][vc_url].
+[vc_url]: http://www.minecraftforum.net/forums/mapping-and-modding/minecraft-mods/2091981-version-checker-auto-update-mods-and-clean
+
 ## 1.2.0
 - Added boilerplate for forge event listeners